Merged STP and STP toolkit notes.
I merged the two files to create a consolidated, one-stop file for STP.
The idea is to make it easier for future study and labs. At this stage
I am unsure if I'll add Rapid STP to this file or create a new one for
it. I'm going to see how much content is there.